+++ module/item/item-magic-sheet.js
@@ -90,6 +90,19 @@ export class ArM5eItemMagicSheet extends ArM5eItemSheet {
     return data;
   }
 
+  static async GetFilteredAspects() {
+    const filterBooks = Object.fromEntries(
+      Object.entries(await game.settings.get(CONFIG.ARM5E.SYSTEM_ID, "sourcebookFilter")).filter(
+        ([key, f]) => f.value === true
+      )
+    );
+    return Object.fromEntries(
+      Object.entries(CONFIG.ARM5E.ASPECTS).filter(([key, val]) => {
+        return val.src in filterBooks;
+      })
+    );
+  }
+
   /* -------------------------------------------- */
 
   // /** @override */