+++ ReleaseNotes.md
@@ -22,6 +22,7 @@
   - Vis costs
   - Migration of legacy magic items
   - Pick the type of enchanted item
+  - Filter shape and material options based on the Sourcebook filter setting
 - [technical] rework and factorization of the foldable sections code
 - [technical] converted larger png assets into webp + userguide update of image links
 - [technical] new framework for lab activities for factorization purposes